Fix bug with dividing on zero in the account details
There is bug in the account detail.
When mongoDB evaluate percentage_field you cannot divide by 0. when last_current or last_past field in the query was 0, mongoDB returns error and server returns 500 Internal Server Error. Now, if last_current or last_past is 0, by default it replace 0 on 1 and all works fine.